abstract status updates into a notification service, notify on connect only if first device went online.
