--- server/src/domains/notifications/services/profileSubscription.ts
@@ -1,3 +1,0 @@
-class ProfileSubscriptionService {
-  constructor() {}
-}
