+++ server/src/domains/users/services/status.ts
@@ -1,16 +1,15 @@
-import type { SelectableUserStatus, UserStatus } from "../interfaces/user.js";
 import { Ok } from "resultat";
-import { wsServerClient } from "../../../index.js";
-import type { ProfileSubscriptionStore } from "../../notifications/index.js";
+import type { NotificationService } from "../../notifications/services/notifications.js";
+import type { SelectableUserStatus, UserStatus } from "../interfaces/user.js";
 import type { IUserModel } from "../models/user.interface.js";
 
 export class StatusService {
   constructor(
     private userModel: IUserModel,
-    private profileSubscriptionStore: ProfileSubscriptionStore
+    private notificationService: NotificationService
   ) {}
 
-  async setStatus(userId: string, status: SelectableUserStatus) {
+  async setStatus(userId: string, selectedStatus: SelectableUserStatus) {
     // Check if status changed to prevent unnecessary notifications.
     // This would be especially bad if user was invisible and went offline.
     const getStatusResult = await this.userModel.getStatus(userId);
@@ -21,33 +20,23 @@ export class StatusService {
 
     const currentStatus = getStatusResult.val;
 
-    if (currentStatus === status) {
+    if (currentStatus === selectedStatus) {
       return Ok(currentStatus);
     }
 
     const updateStatusResult = await this.userModel.updateStatus(
       userId,
-      status
+      selectedStatus
     );
 
     if (!updateStatusResult.ok) {
       return updateStatusResult;
     }
 
-    // TODO: move to a notifcation service
-    const subscribers = this.profileSubscriptionStore.getSubscribers(userId);
+    const status = selectedStatus === "invisible" ? "offline" : selectedStatus;
 
-    const length = subscribers.length;
-    for (let i = 0; i < length; i++) {
-      const subscriberClientId = subscribers[i]!;
-      wsServerClient
-        .toClient(subscriberClientId)
-        .send("subscription:user-profile:status", {
-          profileId: userId,
-          status: status === "invisible" ? "offline" : status,
-        });
-    }
+    this.notificationService.notifyUserProfileStatusChange(userId, status);
 
-    return Ok<UserStatus>(status === "invisible" ? "offline" : status);
+    return Ok<UserStatus>(status);
   }
 }
